Handle the graduation of 12th grades
Handle the case where 12th grades need to be updated, meaning graduated. If the class is 12th grade, this means there's not going to be any grater than 12 so we need to kind of graduate these 12th grades and make them from now on a guest in this server. On the other hand, return something ephemeral so that we don't have problems about non-responded interactions.